// sanity.types.ts
export type SEARCH_SESSIONS_QUERYResult = Array<{
  _id: string;
  startTime: string | null;
  maxCapacity: number | null;
  status: "cancelled" | "completed" | "scheduled" | null;
  activity: {
    _id: string;
    name: string | null;
    instructor: string | null;
    duration: number | null;
    tierLevel: "basic" | "champion" | "performance" | null;
    images: Array<{
      asset: {
        url: string | null;
      } | null;
    }> | null;
  } | null;
  venue: {
    _id: string;
    name: string | null;
    address: {
      fullAddress: string | null;
    } | null;
  } | null;
}>;